+++ topobank/manager/tests/test_publishing.py
@@ -10,12 +10,16 @@ from .utils import SurfaceFactory, UserFactory, Topography2DFactory, TagModelFac
 from ..forms import SurfacePublishForm
 from ..views import SurfaceDetailView
 from topobank.utils import assert_in_content, assert_not_in_content
-from topobank.manager.models import NewPublicationTooFastException, PublicationsDisabledException
+from topobank.manager.models import Surface, NewPublicationTooFastException, PublicationsDisabledException, \
+    PublicationException
+
+# Example user
+bob = dict(first_name='Bob', last_name='Doe', orcid_id='123',
+           affiliations=[dict(name='UofA', ror_id='123')])
 
 
 @pytest.mark.django_db
 def test_publication_version(settings):
-
     settings.MIN_SECONDS_BETWEEN_SAME_SURFACE_PUBLICATIONS = None  # disable
 
     surface = SurfaceFactory()
@@ -38,6 +42,30 @@ def test_publication_disabled(settings):
     with pytest.raises(PublicationsDisabledException):
         surface.publish('cc0-1.0', 'Bob')
 
+
+@pytest.mark.django_db
+def test_publication_superuser(settings):
+    surface = SurfaceFactory()
+    surface.creator.is_superuser = True
+    surface.creator.save()
+    nb_surfaces = len(Surface.objects.all())
+    with pytest.raises(PublicationException):
+        surface.publish('cc0-1.0', 'Bob')
+    # Make sure that the copy, and not the original surface, is deleted again
+    assert len(Surface.objects.all()) == nb_surfaces
+
+
+@pytest.mark.django_db
+def test_failing_publication(settings):
+    settings.DATACITE_API_URL = 'https://nonexistent.api.url/'  # lets publication fail
+    settings.PUBLICATION_DOI_MANDATORY = True  # make sure to contact DataCite
+    settings.PUBLICATION_DOI_PREFIX = '10.12345'
+    surface = SurfaceFactory()
+    nb_surfaces = len(Surface.objects.all())
+    with pytest.raises(PublicationException):
+        surface.publish('cc0-1.0', [bob])
+    # Check that the copy of the surface was properly deleted again
+    assert len(Surface.objects.all()) == nb_surfaces
 
 @pytest.mark.parametrize("should_be_visible", [True, False])
 @pytest.mark.django_db
@@ -135,7 +163,6 @@ def test_permissions_for_published():
 
 @pytest.mark.django_db
 def test_surface_deepcopy():
-
     tag1 = TagModelFactory()
     tag2 = TagModelFactory()
 
@@ -193,7 +220,6 @@ def test_surface_deepcopy():
 
 @pytest.mark.django_db
 def test_switch_versions_on_properties_tab(client, settings, handle_usage_statistics):
-
     settings.MIN_SECONDS_BETWEEN_SAME_SURFACE_PUBLICATIONS = None
 
     user = UserFactory()
@@ -244,7 +270,6 @@ def test_switch_versions_on_properties_tab(client, settings, handle_usage_statis
 
 @pytest.mark.django_db
 def test_notification_saying_new_version_exists(client, settings, handle_usage_statistics, example_authors):
-
     settings.MIN_SECONDS_BETWEEN_SAME_SURFACE_PUBLICATIONS = None
 
     user = UserFactory()
@@ -338,7 +363,6 @@ def test_dont_show_published_surfaces_on_sharing_info(client, example_authors):
 
 @pytest.mark.django_db
 def test_dont_show_published_surfaces_when_shared_filter_used(client, handle_usage_statistics, example_authors):
-
     alice = UserFactory()
     bob = UserFactory()
     surface1 = SurfaceFactory(creator=alice, name="Shared Surface")
@@ -348,11 +372,11 @@ def test_dont_show_published_surfaces_when_shared_filter_used(client, handle_usa
 
     client.force_login(bob)
 
-    response = client.get(reverse('manager:search')+'?sharing_status=shared')  # means "shared with you"
+    response = client.get(reverse('manager:search') + '?sharing_status=shared')  # means "shared with you"
     assert_in_content(response, "Shared Surface")
     assert_not_in_content(response, "Published Surface")
 
-    response = client.get(reverse('manager:search')+'?sharing_status=published')  # means "published by anyone"
+    response = client.get(reverse('manager:search') + '?sharing_status=published')  # means "published by anyone"
     assert_not_in_content(response, "Shared Surface")
     assert_in_content(response, "Published Surface")
 
@@ -449,7 +473,6 @@ def test_publishing_wrong_license(example_authors):
 
 @pytest.mark.django_db
 def test_show_license_and_affiliations_when_viewing_published_surface(rf, settings, example_authors):
-
     license = 'cc0-1.0'
 
     surface = SurfaceFactory()
@@ -470,7 +493,3 @@ def test_show_license_and_affiliations_when_viewing_published_surface(rf, settin
     assert_in_content(response, "Authors")
     assert_in_content(response, request.user.first_name)
     assert_in_content(response, request.user.last_name)
-
-
-
-